Ensure all sever output is printed
Previously anything written to standard out or error was printed at the end of each test, however if the server process didn't flush it's output on printing (which many programming languages don't do by default) the output may not be written until after the the process exits. Make sure we again check if anything is printed after the server process has stopped.

Thanks to Matt Bee for spotting this.